making clone and revival stand alone
